Implement a new parameter to generate only async stubs
This commit adds a new parameter `async_only` to the plugin, that when
present, will only generate the async stubs.

The current {}AsyncStub classes are "fake", they are not present in the
.py file, so they can only be used to write type hints. Using them also
only works if the user manually casts the sync stub to the async stub.

Since most users will only want to use either the sync or the async
stub, this new parameter will allow to generate only the stubs that are
suited for the user's use case, and allow them to have less hacky code.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>